Moved all iTunes elements to their own files
